AE-2168 Fix Gear Down In Transit
Change the order of using the available parameters. We go first for the parameters which
would provide us with the best information.

Updated unit tests to include the case where the gear transitions shortly from Up to
Down. Neither Gear Up nor Gear Down show a change during that time.