AE-2647: Fix radio altitude for go arounds / diversion
Radio altitude was being checked against altitude STD values from the start
and end of the flight. This did not take into account diversions to
airports with a different altitude and resulted in values being subtracted
from rad alt at the airport the flight was origionally intending to land
at. Calculation now uses CCD phases to split the fast phase.