Bugfix/dg/ae 2714 max maintained (#510)
* AE-2714 Fix max_maintained_value

Use an algorithm inspired by second_window to find the largest
maintained value for a given duration.

Fixed the indices in the unittests, as with this new behaviour, the
index will point to the start of the max maintained value period, not to
where the value was observed.

* AE-2714 change max_maintined_value behaviour

When provided with a slice, max_maintained_value will only consider
the slice within the provded array and return the index relative
to the start of the array, not to the start of the slice.

This allowed to simplify the code on the calling site. All KPVs
relying on this function were slightly simplified.