AE-2754 Add all required nodes to the dependency tree node.
No trying to be too clever by optimizing the root edges. We found out
that if a node has a predecessor, its dependencies could not be satisfied
at that time due to circular dependencies.
Trying each node from the root adds slightly more work but assures we
have tried all possible paths to derive the node.