AE-2754 Record cycles seen before to avoid them in the future.
Each cycle encountered is recorded. If the same is seen again, it means
that the nodes in it will never be derivable. So we mark them as
permanently inoperative. This avoids each time to traverse the same cycle
again and again.

It's the case with LFL ATR72_V0 which does not record any Gear positions.
So the cycle Gear Down Selected - Gear Up Selected - Gear Up - Gear Up In Transit -
Gear Down - Gear Down In Transit and Gear In Transit is bypassed.