AE-2657: synchro parameter 4 second window (#490)
* AE-2657: synchro parameter 4 second window

* fix AccelerationLateralOffsetRemoved, add slices_int (#489)

* fix broken EngVibBroadbandMax unit test (#491)

* fix broken EngVibBroadbandMax unit test

* AE-2657: add unit test case

* AE-2684 Fix bug in blend_parameters and process_flight:derive_parameters (#492)

blend_parameters could end up with a slice of length 1 of valid sample, but
might need to shrink it due to different parameter frequencies and the
desired result frequency. This led to a slice to collapse to a 0 length
slice. Now we check for that possibility and discard the empty slice.

In process_flight.py, the derive_parameters function states that the
expected_length for a parameter array should be rounded up. Added the
logic to the code too.

* AE-2657: synchro parameter 4 second window

* AE-2657: add unit test case

Co-authored-by: Daniel Gillet <45759753+gille02d@users.noreply.github.com>