AE-2758 Fix AltitudeAAL by fixing AltitudeRadio (#543)
AltitudeRadio cannot depend on ClimbCruiseDescent. If it does, then
AltitudeAAL will never use AltitudeRadioOffsetRemoved because it depends on
AltitudeRadio -> ClimbCruiseDescent -> Aiborne -> AltitudeAALForFlightPhases
which depends again on AltitudeAAL.

In order to break this cycle, as considering that ClimbCruiseDescent was a nice
idea, the logic for determining ClimbCruiseDescent has been moved to the library.
ClimbCruiseDescent uses it for Airborne slices. But we can also use it for the
whole flight, not bothering with Airborne. This gives a crude idea of where
we climbed, cruised and descended but it's enough for what needs to be done
in AltitudeRadio.