Bugfix/dg/ae 2758 alt aal (#544)
* AE-2758 Fix bug shifting slices while Airborne

* AE-2758 The crude ClimbCruiseDescent will use the Fast section.

As Fast is guaranteed to be available in RadioAltitude, we can use it
to replace Airborne from the ClimbCruiseDescent logic. This will give us
an even better approximation of all the ups and downs, ignoring most of the
ground sections. It hopefully will be good enough for what we need to do
with it here.