AE-2754 Add more nodes to inop_nodes
If the node is not a derived node, we can immediately add it to the
inop_nodes set if it's not present.
For derived nodes, if we could not derive them with all available
dependencies minus the inop nodes, then it will never be possible
to derive it. We can also add it to the inop_nodes set.