AE-2754 Fix bug
When a node cannot operate with the current dependencies, we should
only add it to the inop nodes set if the remaining dependencies based on the
ORIGINAL list of successors minus the inop nodes prevents it from being
operational.
The ordered_successors list could have been modified due to the node
being already present in the dependency path.