AE-2800 Gear Up In Transit extended by 1 sample
When using Gear Up and Gear Down only to determine Gear Up In Transit, if
Gear Up is offset by 0.5 seconds from Gear Down, and as Gear Up In Transit
is aligned to Gear Down, if we align everything to Gear Up, we would find
a gap between Gear Up In Transit and Gear Up.
We decide in this case to extend by 1 sample Gear Up In Transit. Worst case
scenario would be that the Gear Up In Transit still shows Retracting while the
Gear Up shows Up for 1 sample.