AE-2798 Restrict usage of Gear Down Selected to sample rates >= 1Hz
Gear Up Selected should only use Gear Down Selected (from LFL) if
the sample rate is at least 1 Hz. Otherwise we are better off using
the other parameters.

In order to find out if the frequency is sufficient, we need to leave
off the alignment and do the alignment manually.