AE-2754 Use dummy Nodes for testing.
In order to separate the tests from the actual code base, we have extracted
all the Node classes in a test folder, stripping them from unnecessary
methods. This means all the Nodes class have a can_operate method and
an empty derive method with the full signature.

This allows the tests to rely on a constant relationship between the
nodes. Otherwise changes in the dependencies would require to rewrite all the
tests.