AE-2941 Improve nearest_runway algorithm
For non-precise aircrafts, if we have multiple parallel runways, we only
choose the nearest one if the second nearest is at least 150m away from
the touchdown position. Otherwise we fall back to undetermined parallel
runway.

For precise aircrafts, we also choose the nearest runway. If we have multiple
runway less then 30 meters from touchdown position, it must be crossing
runways and we chose the runway with the nearest heading.