AE-2941 Fix compute_track
When calculating LatitudePrepared and LongitudePrepared, we provide the
lat/lon for the liftoff and touchdown. We integrate the speed and heading
to compute the track. But we compute that for an airspeed higher than 50
kt. The liftoff an touchdown points can be significantly away from the
points where airspeed reaches 50 kt.

So now we correctly compute the track based on the liftoff point and
touchdown point, integrating in reverse for the part prior the liftoff
and after the touchdown.