AE-2583 Fine tune Climb Acceleration Start KTI (#453)
* AE-2583 Fine tune Climb Acceleration Start KTI

* Increase moving average from 7 to 11 seconds.
* Filter out airspeed decreases right before flap retraction

* Small fix when checking for speed decrease

* AE-2583 Fix unittest to adapt to the new behaviour.

Now we scan the airspeed backward until we find the moment where the speed
stopped increasing and started to decrease again.

* AE-2583: fixed a typo to trigger build

* AE-2583 Add test case for Airspeed Sel increase after flap retraction

Heavily refactored the derive method to have less nested code.

* AE-2583 Further refinements to the Climb Acceleration Start KTI

- Do not use spd and flap algorithm if the acceleration detected between Flap retraction
  and the moment where airspeed increased is less than 5 kt
- When using airspeed and flap, move the KTI found 5 sec earlier to avoid capturing
  pitch low events.

* AE-2583 Airspeed Selected should not be used if taking off with Flaps 0

* AE-2583 Adjust sensitivity finding airspeed increase before flap retraction

* AE-2583 Parameters fine-tuning.

For detecting start of acceleration using Airspeed, we extend the
slice to see when the flaps were retracted from 4000 to 5000 ft.
Also we only move the detected Climb Acceleration Start by 3 seconds
instead of 5 seconds.

* AE-2583 For B787, use Airspeed Target when available.

On B787, Airspeed Selected is not usefull during the initial climb out.
It seems that Airspeed Target can be used, although it's sometimes showing
zero throughout the flight.

* AE-2583 Better determination of speed selected increase

The previous algorithm could be tricked by small fluctuations. It was
also not behaving properly with different frequencies.

We will now look for a change of at least 5 kt per sample (or less for
high frequency params) and make sure it's staying like that for at least
5 samples

* AE-2583 Use Flap Lever instead of the Flap Lever Set KTI

This makes it easier to figure out if the flaps are set to 0 when Airspeed
Selected increases. All tests adapted accordingly.

Also restrict the use of Airspeed Target to B787.

* AE-2583 Confirm that flaps are retracting

Only look for flap retraction in the given slice. Also make sure the flaps
are indeed retracting and not extending.

* AE-2583 Airspeed Selected increase index earlier by 1 sample

The point of detection for Airspeed Selected increase is now the point
before the speed increased.

* AE-2583 Use Airspeed Sel increase nearest to flap retraction

* AE-2583 Fix issues with slices

Co-authored-by: Michal Mildner <michal.mildner@flightdataservices.com>