Bugfix/dg/ae 2946 repair mask (#604)
* AE-2946 Change repair_mask to copy the array by default

Repairing the array in place is a side-effect which is often used
unknowingly. We should make this a conscious action in the API.

Fixed all points of use in the code where we repair an array and made
sure we use a copy of it where necessary.

* AE-2946 Repair masks for Longitude and Latitude for any number of sources

For consistency, Latitude and Longitude will repair the mask of their sources
with the pre-defined duration (10 sec).

* AE-2946 Make a deepcopy of Latitude parameter

If changing the array in place, we need to make a deepcopy of the parameter.

* AE-2946 Fix FuelQty

We need to keep a copy of the parameter as we need its offset along its array.
Fix except clause which was catching everything and hiding possible bugs.
Fix case where all tanks are masked.
Fix test case which was not failing if a fully masked array was returned.
Added test case for fully masked recorded fuel qty.

* AE-2946 Fix EngStart

At start of recording, engine start is masked out. We have to rely on the
repaired array to determine engine start.

* AE-2946 LandingDecelerationEnd repairs Airspeed

* AE-2946 Fix _smooth_coordinates in CoordinatesStraighten

After repairing the array, we should combine the masks of the repaired arrays.

* AE-2946 Fix Altitude AGL For Flight Phase

Use repaired array to derive Altitude AGL For Flight Phase.

* AE-2946 Use param frequency when repairing mask

* AE-2946 Fix EngStop

When engine is running till the end of the recording, the signal drops
suddenly and gets masked. Small portions could still be unmasked with
values around 0.

We need to find the last valid sample with engine running and select the
next sample as the Eng Stop KTI.

* AE-2946 Fix EngStart KTI

When Engine has started before the recording, there could be small unmasked values
at the start of recording. We want to detect the eng start as the sample prior to the
first valid sample of running engines.

* AE-2946 Fix various Airspeed KPVs bewteen altitude bands

We only want to trigger one KPV per altitude band, even if there are masked
sections in that band.

* AE-2946 Flap Angle properly handles masked values with more than 2 sources

With 4 sources, if all the sources have the same offset, we average the values which
allows to ignore any masked values, provided at least one source has valid data.

With 4 sources at different offsets, we repair any masked values provided at least
one other source has valid data. Then we repair 10 sec of data with the fill_start
method.

* AE-2946 Fix index_at_value for closing or first_closing endpoint

Fixed backward slices indexes. If we want to select all the elements backwards
up to the first one included we must use None as the end index in the slice:
slice(some_value, None, -1)
With masked values, we want first to repair the array so we can find the best
index estimate where the value should have crossed the threshold. If it never
crosses the threshold, it has the nice consequence that we can now find easily
the point where it was last closing in.

Fixed consequently Landing phase.
Fixed step_values as it was using a reversed slice assuming the end of the slice
would be included in the search, but it's always excluded.

* AE-2946 Fix TouchdownToPitch2DegreesAbovePitchAt60KtsDuration

Fix checking if pitch value found 60 kts is None. It can be 0.0.
Also fixed how we find Airspeed at 60 kts. In Parameter Operating Limits,
Airspeed lower limit is set at 60kt. So we often miss it. But we cannot rely
on the values that have been masked as they could be corrupted. So we use
the masked array and find the nearest point to 60 kts as our reference.

* AW-2946 Fix Flap Lever Synthetic mask

If Flap array has masked values, they should be propagated to the Flap Lever
Synthetic array.