Bugfix/dg/ae 2914 go around gear up (#594)
* AE-2914 GearUpSelectionDuringGoAround scans up to the next Top Of Climb

The Gear could be raised after the Go Around and Climbout phase. So we extend
the period we are looking for a Gear Up Selection up to the following Top
Of Climb.
If we don't have a Top Of Climb, we revert to the Go Around And Climbout
phase.

* AE-2914 AltitudeAtGearUpSelectionDuringGoAround reworked

We now scan from the lowest altitude point until the following top of climb
for a Gear Up Selection During Go Around KTI. Without a following TOC, we
scan within the Go Around and Climbout section.

No KPV will be generated without a Gear Up Selection During Go Around KTI.

* AE-2914 Add new KPV Gear Down Selected Duration After Go Around

This new KPV will measure the time it took to raise the Gear after a Go Around.
It will scan for Gear Up Selection up until the following Top Of Climb. If no
Gear Up was obserbed, it will return the total duration until reaching the Top
Of Climb. This will allow to detect go arounds where the gear was not retracted.

* AE-2914 Fix Descent Low Climb phase.

This phase should be created when Altitude AAL dips lower than 3,000ft.
This is now in line with the Go Around And Climbout phase.

* AE-2914 Cover unlikely edge case

* AE-2914 Rename KPV to conform to naming convention

* AE-2914 Improve code readability using ternary operator.