Bugfix/dg/ae 2825 eng running (#570)
* AE-2825 Ignore masked N2 or FF data.

When using N2 and/or FF, only use the non-masked data.

* AE-2825 Avoid deprecation warning

DeprecationWarning: Calling nonzero on 0d arrays is deprecated, as it behaves
surprisingly. Use `atleast_1d(cond).nonzero()` if the old behavior was
intended. If the context of this warning is of the form `arr[nonzero(cond)]`,
just use `arr[cond]`.