AE-2962 Fix some bugs introced by the new slices_or function
There is one test case with a huge amount of slices. I've made a graph of it
on matplotlib and visually compared the all those slices spread over a single line
with the function result. After these changes, the expected result is now correct.

Added a few more test cases to cover some edge cases.