AE-2962 Fix slices_extend
slices_extend can only work on slices with positive indices and a step of 1.
There was a test case for one slice with a step of 3, but if we are adding a
second slice overlapping the first with a step of 3, the result was wrong.

We now check for that restriction.

Factored out the check by creating a new function: are_slices_straight_continuous
Using this also in slices_or function.

Also slices_extend should not change a starting None into a 0. We now preserve that.