AE-2941 Use short slices of lat / lon to determine runways
For the landing runway, we use the lat / lon from 100 ft until 0 ft if available.
We compute the average cross-track distance to each runway. If the aircraft has
non-precise positioning and if 2 or more parallel runways are within 135 meters
of the cross-track distance, we default to an undefined RWY.

Taking a slice allows for a better estimate of the runway than a single point.

For takeoff we do something similar, taking a lat / lon slice from acceleration
start until liftoff if there are available.

Unittests added plus some additional test cases with real data.