AE-2973 Fix Airspeed Relative For 3 Sec
During regression tests for release 20200922, it was found out that 2 KPVs were
missing on 119 flights:

Airspeed Relative For 3 Sec 20 Ft To Touchdown Max
Airspeed Relative For 3 Sec 20 Ft To Touchdown Min

A short investigation revealed that those KPVs first trim the slice from 20 Ft
to Touchdown by 3 seconds on both sides. This means that if the duration
between 20 Ft and Touchdown is less than 6 seconds, the slice won't exist
anymore and the KPV will not trigger.

Removed the trimming from Airspeed Relative For 3 Sec Min/Max at the following
heights:
  - 20 Ft To Touchdown
  - 500 Ft To 20 Ft
  - 1000 Ft To 500 Ft