AE-2987 Create Initial Climb and Climb phases after Descent
If Helicopter descends but climbs again, we now create a new Initial Climb
up to 1,000 ft and then a Climb. This means that a ClimbCruiseDescent is
now filled with either an Initial Climb, a Climb, a Cruise or a Descent.

Altitude AAL for helicopters consider altitude changes of 50 ft instead of
500 ft for airplanes.

We repair Gear On Ground for helicopter Airborne flight phase.