AE-2995 Takeoff cannot start in previous Fast section
During touch and go's, if Airspeed was low for a significant amount of time,
the Fast section would end and a new one would start slightly later.
This means that a new takeoff would be determine. But it should be restricted
to start at earliest at the end of the previous Fast section.